Remove the class PackagePayload
Merge the class PackagePayload with the class DNFPayload.
