+++ pyanaconda/ui/gui/spokes/software_selection.py
@@ -23,7 +23,7 @@ import gi
 from pyanaconda.flags import flags
 from pyanaconda.core.i18n import _, C_, CN_
 from pyanaconda.core.configuration.anaconda import conf
-from pyanaconda.payload.base import PackagePayload
+from pyanaconda.payload.dnf import DNFPayload
 from pyanaconda.payload.manager import payloadMgr, PayloadState
 from pyanaconda.payload.errors import NoSuchGroup, PayloadError, DependencyError
 from pyanaconda.threading import threadMgr, AnacondaThread
@@ -289,7 +289,7 @@ class SoftwareSelectionSpoke(NormalSpoke):
 
     @property
     def showable(self):
-        return isinstance(self.payload, PackagePayload)
+        return isinstance(self.payload, DNFPayload)
 
     @property
     def status(self):
