+++ pyanaconda/ui/tui/spokes/installation_source.py
@@ -23,7 +23,7 @@ from pyanaconda.ui.categories.software import SoftwareCategory
 from pyanaconda.ui.tui.spokes import NormalTUISpoke
 from pyanaconda.ui.tui.tuiobject import Dialog
 from pyanaconda.threading import threadMgr, AnacondaThread
-from pyanaconda.payload.base import PackagePayload
+from pyanaconda.payload.dnf import DNFPayload
 from pyanaconda.payload import utils as payload_utils
 from pyanaconda.payload.manager import payloadMgr, PayloadState
 from pyanaconda.core.i18n import N_, _, C_
@@ -126,7 +126,7 @@ class SourceSpoke(NormalTUISpoke, SourceSwitchHandler):
 
     @property
     def showable(self):
-        return isinstance(self.payload, PackagePayload)
+        return isinstance(self.payload, DNFPayload)
 
     @property
     def status(self):
